uint64_t -> size_t / int64_t -> ssize_t

+++ include/tmc/asio/aw_asio.hpp
@@ -107,7 +107,7 @@ template <IsAwAsio Awaitable> struct awaitable_traits<Awaitable> {
     awaitable.customizer.done_count = DoneCount;
   }
 
-  static void set_flags(self_type& awaitable, uint64_t Flags) {
+  static void set_flags(self_type& awaitable, size_t Flags) {
     awaitable.customizer.flags = Flags;
   }
 };
